refactor : [TSK-20] firebase app 초기화(CRA=>vite 마이그레이션)
